add testing for check wining conditions
