+++ src/Players.js
@@ -1,0 +1,4 @@
+const Player = (name, mark, score) => {
+  return { name, mark, score };
+};
+export default Player;
